requirements.txt
django==2.1.3
django-crispy-forms==1.7.2
django-ckeditor==5.6.1